+++ NEWS.md
@@ -3,6 +3,7 @@
  * Maths was corrupted if the window was temporarily not wide
    enough to display it in 2D (#1812)
  * Correct label scaling after changing the zoom factor (#1815)
+ * Repaired "copy as bitmap" (#1820)
 
 # 23.10.0
  * A big rewrite to make the static analysis tools happy